pyproject.toml
[project]
name = "quantmind"
version = "0.1.0"
description = "A production-grade, fully open-source multi-agent investment research system."
requires-python = ">=3.10"
dependencies = [
    "langchain",
    "langchain-groq",
    "langchain-google-genai",
    "langchain-ollama",
    "langchain-openai",
    "langgraph",
    "deepagents",
    "pydantic",
    "pydantic-settings",
    "httpx",
    "tenacity",
    "python-dotenv",
    "fastapi",
    "uvicorn",
    "streamlit",
    "yfinance",
    "sec-edgar-downloader",
    "fredapi",
    "praw",
    "requests",
    "newsapi-python",
    "transformers",
    "torch",
    "sentence-transformers",
    "chromadb",
    "duckduckgo-search",
    "prophet",
    "statsmodels",
    "pandas",
    "numpy",
    "scipy",
    "diskcache",
    "langfuse",
    "pytest",
    "pytest-asyncio",
    "mlx-lm; sys_platform == 'darwin'",
    "huggingface-hub",
    "arize-phoenix",
    "openinference-instrumentation-langchain",
    "fastmcp"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
